Added basic level save and load functionality
- Created Binary type to make copying data to a char* format easier
- Added base node seralisation function, this will store transform along with other base node variables
- This function can be overriden if we want to save things on the object (such as sprite textures, by default constructor makes sprites blank, so need to store them)
- More Engine node save functions to be added later
- Added TypeRegsitry class to the Game project side of things. This should allow users to add their own custom node types to be spawned on level load.
- Users will have to define a uint32_t type id when a node is constructed using the base node "Type" variable
- Currently there are a lot of steps required to actually get a node hooked up to serialization. However I imagine most nodes will be fine being spawned in with their default parameters in the positions they were when the level saved
- Currently renderer is not flushed when level is changed
- Currently level only saves and loads from on location
- NOTE: ONLY HEAP ALLOCATED NODES CAN BE SAVED AND LOADED RIGHT NOW. I'd imagine most would want to spawn nodes with self managed lifetimes but just to note.

Extra:
- For fun added a meshrenderer node type to see if I could. Turns out I could pretty easily.

+++ ChaosEngine/src/Chaos/DataTypes/ChaoticArray.h
@@ -21,7 +21,7 @@ namespace Chaos
 			Data = (T*)malloc(size * sizeof(T));
 			memset(Data, 0, size * sizeof(T));
 			
-			FreeSlots = (int*)malloc(size * sizeof(int));
+			FreeSlots = (size_t*)malloc(size * sizeof(size_t));
 			//memset(FreeSlots, 0, size * sizeof(int));
 			
 			for (size_t i = 0; i < size; ++i)
@@ -136,7 +136,7 @@ namespace Chaos
 		size_t FreeSlotsSize() { return m_freeSlotsSize; }
 		
 		T* Data = nullptr;
-		int* FreeSlots = nullptr;
+		size_t* FreeSlots = nullptr;
 		
 		private:
 		size_t m_size = 0;
